vendor an esbuild output version of the typescript package to work around run time issues
I cloned down the `typescript` package and ran

npx esbuild --bundle --platform=browser --target=node22 --format=esm  --outfile=./BIG_TS.js src/typescript/typescript.ts

this output a slightly different compatibility wrapper that correctly
produces a module with no default exports, which allows the build time
and runtime values to be populated.

It will fail TSC, because we have not type definition file, but we could
grab that from the typescript project if we really wanted to.

This is probably not a good solution, but it points to the core issue
being the export wrapper of `tsc` and whatever is happening in the
component test rollup transforms.